Move FizzBuzz class to it project folder and java package
